+++ tests/test_polars.py
@@ -2,7 +2,10 @@ import pytest
 
 from itables import to_html_datatable
 from itables.javascript import datatables_rows
-from itables.sample_dfs import get_dict_of_test_dfs, get_dict_of_test_series
+from itables.sample_dfs import (
+    get_dict_of_polars_test_dfs,
+    get_dict_of_polars_test_series,
+)
 
 try:
     import polars as pl  # noqa
@@ -11,14 +14,14 @@ except ImportError as e:
 
 
 @pytest.mark.parametrize(
-    "name,x", [(name, x) for name, x in get_dict_of_test_series(polars=True).items()]
+    "name,x", [(name, x) for name, x in get_dict_of_polars_test_series().items()]
 )
 def test_show_polars_series(name, x):
     to_html_datatable(x)
 
 
 @pytest.mark.parametrize(
-    "name,df", [(name, df) for name, df in get_dict_of_test_dfs(polars=True).items()]
+    "name,df", [(name, df) for name, df in get_dict_of_polars_test_dfs().items()]
 )
 def test_show_polars_df(name, df):
     to_html_datatable(df)