+++ docs/changelog.md
@@ -10,6 +10,7 @@ ITables ChangeLog
 **Changed**
 - The `warn_on_undocumented_option` option now defaults to `True` and only check the option names
 - We have added a new option `warn_on_unexpected_option_type` that defaults to `warn_on_undocumented_option` when `typeguard>=4.4.1` is installed (`False` otherwise)
+- Wrapping JS function definitions in parentheses prior to JS eval is now done on the JS side.
 
 **Fixed**
 - We have added type hints to `itable.options` even for the options that don't have a default value ([#224](https://github.com/mwouts/itables/issues/224))